split the grpc handlers
