+++ danube-broker/src/broker_server/discovery_handler.rs
@@ -1,0 +1,99 @@
+use crate::broker_server::DanubeServerImpl;
+use crate::{broker_service::validate_topic, error_message::create_error_status};
+
+use crate::proto::{
+    discovery_server::Discovery, topic_lookup_response::LookupType, ErrorType, SchemaRequest,
+    SchemaResponse, TopicLookupRequest, TopicLookupResponse,
+};
+
+use tonic::{Code, Request, Response};
+use tracing::{trace, Level};
+
+#[tonic::async_trait]
+impl Discovery for DanubeServerImpl {
+    // finds topic to broker assignment
+    #[tracing::instrument(level = Level::INFO, skip_all)]
+    async fn topic_lookup(
+        &self,
+        request: Request<TopicLookupRequest>,
+    ) -> std::result::Result<Response<TopicLookupResponse>, tonic::Status> {
+        let req = request.into_inner();
+
+        trace!("Topic Lookup request for topic: {}", req.topic);
+
+        // The topic format is /{namespace_name}/{topic_name}
+        if !validate_topic(&req.topic) {
+            let error_string =
+                "The topic has an invalid format, should be: /namespace_name/topic_name";
+            let status = create_error_status(
+                Code::InvalidArgument,
+                ErrorType::InvalidTopicName,
+                error_string,
+                None,
+            );
+            return Err(status);
+        }
+
+        let service = self.service.lock().await;
+
+        let result = match service.lookup_topic(&req.topic).await {
+            Some((true, _)) => (self.broker_addr.to_string(), LookupType::Connect),
+            Some((false, addr)) => (addr, LookupType::Redirect),
+            None => {
+                let error_string = &format!("Unable to find the requested topic: {}", &req.topic);
+                let status = create_error_status(
+                    Code::InvalidArgument,
+                    ErrorType::TopicNotFound,
+                    error_string,
+                    None,
+                );
+                return Err(status);
+            }
+        };
+
+        let response = TopicLookupResponse {
+            request_id: req.request_id,
+            response_type: result.1.into(),
+            broker_service_url: result.0,
+        };
+
+        Ok(tonic::Response::new(response))
+    }
+    // Retrieve message schema from Metadata Store
+    #[tracing::instrument(level = Level::INFO, skip_all)]
+    async fn get_schema(
+        &self,
+        request: Request<SchemaRequest>,
+    ) -> std::result::Result<Response<SchemaResponse>, tonic::Status> {
+        let req = request.into_inner();
+
+        trace!("Schema Lookup for the topic: {}", req.topic);
+
+        // The topic format is /{namespace_name}/{topic_name}
+        if !validate_topic(&req.topic) {
+            let error_string =
+                "The topic has an invalid format, should be: /namespace_name/topic_name";
+            let status = create_error_status(
+                Code::InvalidArgument,
+                ErrorType::InvalidTopicName,
+                error_string,
+                None,
+            );
+            return Err(status);
+        }
+
+        let service = self.service.lock().await;
+
+        let proto_schema = service.get_schema(&req.topic);
+
+        // should I inform the client that the topic is not served by this broker ?
+        // as the get_schema is local to this broker
+
+        let response = SchemaResponse {
+            request_id: req.request_id,
+            schema: proto_schema,
+        };
+
+        Ok(tonic::Response::new(response))
+    }
+}
